Add Release workflow on master branch
I detected that this workflow only exists on the 'helmrepo'
branch. When there is a merge between 'master' and 'helmrepo' this is
the main difference between the two branches besides the helm chart
version.

It is save to push this in master branch and reduce the difference
between branches, because the workflow already filters by branch.