make node image configurable for evp
Two different images can be used as node component. The image from
Thingsboard or the one coming from EVP.

We only want to allow replacements of the image of the node component.

The file evp-node.yaml  is added to be updated with any setup related to
evp image without having to touch the one from thingsboard, minimizing
any issue just because the same file is updated.

The images will be set up in evp-chart and not in this repository
because this is public.

Change-Id: Id019f6d80d4cd6bda5ac072756f17fb5014547d1